Make tests pass, allow struct or *struct to retrieve node key
