+++ server/service/endpoint_osquery.go
@@ -37,34 +37,6 @@ func makeEnrollAgentEndpoint(svc fleet.Service) endpoint.Endpoint {
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-// Write Distributed Query Results
-////////////////////////////////////////////////////////////////////////////////
-
-type SubmitDistributedQueryResultsRequest struct {
-	NodeKey  string                               `json:"node_key"`
-	Results  fleet.OsqueryDistributedQueryResults `json:"queries"`
-	Statuses map[string]fleet.OsqueryStatus       `json:"statuses"`
-	Messages map[string]string                    `json:"messages"`
-}
-
-type submitDistributedQueryResultsResponse struct {
-	Err error `json:"error,omitempty"`
-}
-
-func (r submitDistributedQueryResultsResponse) error() error { return r.Err }
-
-func makeSubmitDistributedQueryResultsEndpoint(svc fleet.Service) endpoint.Endpoint {
-	return func(ctx context.Context, request interface{}) (interface{}, error) {
-		req := request.(SubmitDistributedQueryResultsRequest)
-		err := svc.SubmitDistributedQueryResults(ctx, req.Results, req.Statuses, req.Messages)
-		if err != nil {
-			return submitDistributedQueryResultsResponse{Err: err}, nil
-		}
-		return submitDistributedQueryResultsResponse{}, nil
-	}
-}
-
-////////////////////////////////////////////////////////////////////////////////
 // Submit Logs
 ////////////////////////////////////////////////////////////////////////////////
