+++ server/service/osquery.go
@@ -3,12 +3,21 @@ package service
 import (
 	"context"
 	"encoding/json"
+	"errors"
+	"fmt"
+	"regexp"
+	"strconv"
+	"strings"
 	"time"
 
 	"github.com/fleetdm/fleet/v4/server/contexts/ctxerr"
 	hostctx "github.com/fleetdm/fleet/v4/server/contexts/host"
+	"github.com/fleetdm/fleet/v4/server/contexts/logging"
 	"github.com/fleetdm/fleet/v4/server/fleet"
+	"github.com/fleetdm/fleet/v4/server/ptr"
+	"github.com/fleetdm/fleet/v4/server/pubsub"
 	"github.com/fleetdm/fleet/v4/server/service/osquery_utils"
+	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
 	"github.com/spf13/cast"
 )
@@ -345,4 +354,447 @@ func (svc *Service) policyQueriesForHost(ctx context.Context, host *fleet.Host) 
 		return nil, ctxerr.Wrap(ctx, err, "retrieve policy queries")
 	}
 	return policyQueries, nil
+}
+
+////////////////////////////////////////////////////////////////////////////////
+// Write Distributed Query Results
+////////////////////////////////////////////////////////////////////////////////
+
+// When a distributed query has no results, the JSON schema is
+// inconsistent, so we use this shim and massage into a consistent
+// schema. For example (simplified from actual osqueryd 1.8.2 output):
+// {
+// "queries": {
+//   "query_with_no_results": "", // <- Note string instead of array
+//   "query_with_results": [{"foo":"bar","baz":"bang"}]
+//  },
+// "node_key":"IGXCXknWQ1baTa8TZ6rF3kAPZ4\/aTsui"
+// }
+type submitDistributedQueryResultsRequestShim struct {
+	NodeKey  string                     `json:"node_key"`
+	Results  map[string]json.RawMessage `json:"queries"`
+	Statuses map[string]interface{}     `json:"statuses"`
+	Messages map[string]string          `json:"messages"`
+}
+
+func (shim *submitDistributedQueryResultsRequestShim) toRequest(ctx context.Context) (*SubmitDistributedQueryResultsRequest, error) {
+	results := fleet.OsqueryDistributedQueryResults{}
+	for query, raw := range shim.Results {
+		queryResults := []map[string]string{}
+		// No need to handle error because the empty array is what we
+		// want if there was an error parsing the JSON (the error
+		// indicates that osquery sent us incosistently schemaed JSON)
+		_ = json.Unmarshal(raw, &queryResults)
+		results[query] = queryResults
+	}
+
+	// Statuses were represented by strings in osquery < 3.0 and now
+	// integers in osquery > 3.0. Massage to string for compatibility with
+	// the service definition.
+	statuses := map[string]fleet.OsqueryStatus{}
+	for query, status := range shim.Statuses {
+		switch s := status.(type) {
+		case string:
+			sint, err := strconv.Atoi(s)
+			if err != nil {
+				return nil, ctxerr.Wrap(ctx, err, "parse status to int")
+			}
+			statuses[query] = fleet.OsqueryStatus(sint)
+		case float64:
+			statuses[query] = fleet.OsqueryStatus(s)
+		default:
+			return nil, ctxerr.Errorf(ctx, "query status should be string or number, got %T", s)
+		}
+	}
+
+	return &SubmitDistributedQueryResultsRequest{
+		NodeKey:  shim.NodeKey,
+		Results:  results,
+		Statuses: statuses,
+		Messages: shim.Messages,
+	}, nil
+}
+
+type SubmitDistributedQueryResultsRequest struct {
+	NodeKey  string                               `json:"node_key"`
+	Results  fleet.OsqueryDistributedQueryResults `json:"queries"`
+	Statuses map[string]fleet.OsqueryStatus       `json:"statuses"`
+	Messages map[string]string                    `json:"messages"`
+}
+
+type submitDistributedQueryResultsResponse struct {
+	Err error `json:"error,omitempty"`
+}
+
+func (r submitDistributedQueryResultsResponse) error() error { return r.Err }
+
+func submitDistributedQueryResultsEndpoint(ctx context.Context, request interface{}, svc fleet.Service) (interface{}, error) {
+	shim := request.(*submitDistributedQueryResultsRequestShim)
+	req, err := shim.toRequest(ctx)
+	if err != nil {
+		fmt.Println(">>>>> shimtorequest: ", err)
+		return submitDistributedQueryResultsResponse{Err: err}, nil
+	}
+
+	err = svc.SubmitDistributedQueryResults(ctx, req.Results, req.Statuses, req.Messages)
+	if err != nil {
+		fmt.Println(">>>>> submitdqr: ", err)
+		return submitDistributedQueryResultsResponse{Err: err}, nil
+	}
+	return submitDistributedQueryResultsResponse{}, nil
+}
+
+const (
+	// hostLabelQueryPrefix is appended before the query name when a query is
+	// provided as a label query. This allows the results to be retrieved when
+	// osqueryd writes the distributed query results.
+	hostLabelQueryPrefix = "fleet_label_query_"
+
+	// hostDetailQueryPrefix is appended before the query name when a query is
+	// provided as a detail query.
+	hostDetailQueryPrefix = "fleet_detail_query_"
+
+	// hostAdditionalQueryPrefix is appended before the query name when a query is
+	// provided as an additional query (additional info for hosts to retrieve).
+	hostAdditionalQueryPrefix = "fleet_additional_query_"
+
+	// hostPolicyQueryPrefix is appended before the query name when a query is
+	// provided as a policy query. This allows the results to be retrieved when
+	// osqueryd writes the distributed query results.
+	hostPolicyQueryPrefix = "fleet_policy_query_"
+
+	// hostDistributedQueryPrefix is appended before the query name when a query is
+	// run from a distributed query campaign
+	hostDistributedQueryPrefix = "fleet_distributed_query_"
+)
+
+func (svc *Service) SubmitDistributedQueryResults(
+	ctx context.Context,
+	results fleet.OsqueryDistributedQueryResults,
+	statuses map[string]fleet.OsqueryStatus,
+	messages map[string]string,
+) error {
+	// skipauth: Authorization is currently for user endpoints only.
+	svc.authz.SkipAuthorization(ctx)
+
+	host, ok := hostctx.FromContext(ctx)
+	if !ok {
+		return osqueryError{message: "internal error: missing host from request context"}
+	}
+
+	detailUpdated := false
+	additionalResults := make(fleet.OsqueryDistributedQueryResults)
+	additionalUpdated := false
+	labelResults := map[uint]*bool{}
+	policyResults := map[uint]*bool{}
+
+	svc.maybeDebugHost(ctx, host, results, statuses, messages)
+
+	for query, rows := range results {
+		// osquery docs say any nonzero (string) value for status indicates a query error
+		status, ok := statuses[query]
+		failed := ok && status != fleet.StatusOK
+		if failed && messages[query] != "" && !noSuchTableRegexp.MatchString(messages[query]) {
+			level.Debug(svc.logger).Log("query", query, "message", messages[query])
+		}
+		var err error
+		switch {
+		case strings.HasPrefix(query, hostDetailQueryPrefix):
+			trimmedQuery := strings.TrimPrefix(query, hostDetailQueryPrefix)
+			var ingested bool
+			ingested, err = svc.directIngestDetailQuery(ctx, host, trimmedQuery, rows, failed)
+			if !ingested && err == nil {
+				err = svc.ingestDetailQuery(ctx, host, trimmedQuery, rows)
+				// No err != nil check here because ingestDetailQuery could have updated
+				// successfully some values of host.
+				detailUpdated = true
+			}
+		case strings.HasPrefix(query, hostAdditionalQueryPrefix):
+			name := strings.TrimPrefix(query, hostAdditionalQueryPrefix)
+			additionalResults[name] = rows
+			additionalUpdated = true
+		case strings.HasPrefix(query, hostLabelQueryPrefix):
+			err = ingestMembershipQuery(hostLabelQueryPrefix, query, rows, labelResults, failed)
+		case strings.HasPrefix(query, hostPolicyQueryPrefix):
+			err = ingestMembershipQuery(hostPolicyQueryPrefix, query, rows, policyResults, failed)
+		case strings.HasPrefix(query, hostDistributedQueryPrefix):
+			err = svc.ingestDistributedQuery(ctx, *host, query, rows, failed, messages[query])
+		default:
+			err = osqueryError{message: "unknown query prefix: " + query}
+		}
+
+		if err != nil {
+			logging.WithErr(ctx, ctxerr.New(ctx, "error in query ingestion"))
+			logging.WithExtras(ctx, "ingestion-err", err)
+		}
+	}
+
+	ac, err := svc.ds.AppConfig(ctx)
+	if err != nil {
+		return ctxerr.Wrap(ctx, err, "getting app config")
+	}
+
+	if len(labelResults) > 0 {
+		if err := svc.task.RecordLabelQueryExecutions(ctx, host, labelResults, svc.clock.Now(), ac.ServerSettings.DeferredSaveHost); err != nil {
+			logging.WithErr(ctx, err)
+		}
+	}
+
+	if len(policyResults) > 0 {
+		if ac.WebhookSettings.FailingPoliciesWebhook.Enable {
+			incomingResults := filterPolicyResults(policyResults, ac.WebhookSettings.FailingPoliciesWebhook.PolicyIDs)
+			if failingPolicies, passingPolicies, err := svc.ds.FlippingPoliciesForHost(ctx, host.ID, incomingResults); err != nil {
+				logging.WithErr(ctx, err)
+			} else {
+				// Register the flipped policies on a goroutine to not block the hosts on redis requests.
+				go func() {
+					if err := svc.registerFlippedPolicies(ctx, host.ID, host.Hostname, failingPolicies, passingPolicies); err != nil {
+						logging.WithErr(ctx, err)
+					}
+				}()
+			}
+		}
+		// NOTE(mna): currently, failing policies webhook wouldn't see the new
+		// flipped policies on the next run if async processing is enabled and the
+		// collection has not been done yet (not persisted in mysql). Should
+		// FlippingPoliciesForHost take pending redis data into consideration, or
+		// maybe we should impose restrictions between async collection interval
+		// and policy update interval?
+
+		if err := svc.task.RecordPolicyQueryExecutions(ctx, host, policyResults, svc.clock.Now(), ac.ServerSettings.DeferredSaveHost); err != nil {
+			logging.WithErr(ctx, err)
+		}
+	}
+
+	if additionalUpdated {
+		additionalJSON, err := json.Marshal(additionalResults)
+		if err != nil {
+			logging.WithErr(ctx, err)
+		} else {
+			additional := json.RawMessage(additionalJSON)
+			if err := svc.ds.SaveHostAdditional(ctx, host.ID, &additional); err != nil {
+				logging.WithErr(ctx, err)
+			}
+		}
+	}
+
+	if detailUpdated {
+		host.DetailUpdatedAt = svc.clock.Now()
+	}
+
+	refetchRequested := host.RefetchRequested
+	if refetchRequested {
+		host.RefetchRequested = false
+	}
+
+	if refetchRequested || detailUpdated {
+		appConfig, err := svc.ds.AppConfig(ctx)
+		if err != nil {
+			logging.WithErr(ctx, err)
+		} else {
+			if appConfig.ServerSettings.DeferredSaveHost {
+				go svc.serialUpdateHost(host)
+			} else {
+				if err := svc.ds.UpdateHost(ctx, host); err != nil {
+					logging.WithErr(ctx, err)
+				}
+			}
+		}
+	}
+
+	return nil
+}
+
+var noSuchTableRegexp = regexp.MustCompile(`^no such table: \S+$`)
+
+func (svc *Service) directIngestDetailQuery(ctx context.Context, host *fleet.Host, name string, rows []map[string]string, failed bool) (ingested bool, err error) {
+	config, err := svc.ds.AppConfig(ctx)
+	if err != nil {
+		return false, osqueryError{message: "ingest detail query: " + err.Error()}
+	}
+
+	detailQueries := osquery_utils.GetDetailQueries(config, svc.config)
+	query, ok := detailQueries[name]
+	if !ok {
+		return false, osqueryError{message: "unknown detail query " + name}
+	}
+	if query.DirectIngestFunc != nil {
+		err = query.DirectIngestFunc(ctx, svc.logger, host, svc.ds, rows, failed)
+		if err != nil {
+			return false, osqueryError{
+				message: fmt.Sprintf("ingesting query %s: %s", name, err.Error()),
+			}
+		}
+		return true, nil
+	}
+	return false, nil
+}
+
+// ingestDistributedQuery takes the results of a distributed query and modifies the
+// provided fleet.Host appropriately.
+func (svc *Service) ingestDistributedQuery(ctx context.Context, host fleet.Host, name string, rows []map[string]string, failed bool, errMsg string) error {
+	trimmedQuery := strings.TrimPrefix(name, hostDistributedQueryPrefix)
+
+	campaignID, err := strconv.Atoi(osquery_utils.EmptyToZero(trimmedQuery))
+	if err != nil {
+		return osqueryError{message: "unable to parse campaign ID: " + trimmedQuery}
+	}
+
+	// Write the results to the pubsub store
+	res := fleet.DistributedQueryResult{
+		DistributedQueryCampaignID: uint(campaignID),
+		Host:                       host,
+		Rows:                       rows,
+	}
+	if failed {
+		res.Error = &errMsg
+	}
+
+	err = svc.resultStore.WriteResult(res)
+	if err != nil {
+		var pse pubsub.Error
+		ok := errors.As(err, &pse)
+		if !ok || !pse.NoSubscriber() {
+			return osqueryError{message: "writing results: " + err.Error()}
+		}
+
+		// If there are no subscribers, the campaign is "orphaned"
+		// and should be closed so that we don't continue trying to
+		// execute that query when we can't write to any subscriber
+		campaign, err := svc.ds.DistributedQueryCampaign(ctx, uint(campaignID))
+		if err != nil {
+			if err := svc.liveQueryStore.StopQuery(strconv.Itoa(campaignID)); err != nil {
+				return osqueryError{message: "stop orphaned campaign after load failure: " + err.Error()}
+			}
+			return osqueryError{message: "loading orphaned campaign: " + err.Error()}
+		}
+
+		if campaign.CreatedAt.After(svc.clock.Now().Add(-1 * time.Minute)) {
+			// Give the client a minute to connect before considering the
+			// campaign orphaned
+			return osqueryError{message: "campaign waiting for listener (please retry)"}
+		}
+
+		if campaign.Status != fleet.QueryComplete {
+			campaign.Status = fleet.QueryComplete
+			if err := svc.ds.SaveDistributedQueryCampaign(ctx, campaign); err != nil {
+				return osqueryError{message: "closing orphaned campaign: " + err.Error()}
+			}
+		}
+
+		if err := svc.liveQueryStore.StopQuery(strconv.Itoa(campaignID)); err != nil {
+			return osqueryError{message: "stopping orphaned campaign: " + err.Error()}
+		}
+
+		// No need to record query completion in this case
+		return osqueryError{message: "campaign stopped"}
+	}
+
+	err = svc.liveQueryStore.QueryCompletedByHost(strconv.Itoa(campaignID), host.ID)
+	if err != nil {
+		return osqueryError{message: "record query completion: " + err.Error()}
+	}
+
+	return nil
+}
+
+// ingestMembershipQuery records the results of label queries run by a host
+func ingestMembershipQuery(
+	prefix string,
+	query string,
+	rows []map[string]string,
+	results map[uint]*bool,
+	failed bool,
+) error {
+	trimmedQuery := strings.TrimPrefix(query, prefix)
+	trimmedQueryNum, err := strconv.Atoi(osquery_utils.EmptyToZero(trimmedQuery))
+	if err != nil {
+		return fmt.Errorf("converting query from string to int: %w", err)
+	}
+	// A label/policy query matches if there is at least one result for that
+	// query. We must also store negative results.
+	if failed {
+		results[uint(trimmedQueryNum)] = nil
+	} else {
+		results[uint(trimmedQueryNum)] = ptr.Bool(len(rows) > 0)
+	}
+
+	return nil
+}
+
+// ingestDetailQuery takes the results of a detail query and modifies the
+// provided fleet.Host appropriately.
+func (svc *Service) ingestDetailQuery(ctx context.Context, host *fleet.Host, name string, rows []map[string]string) error {
+	config, err := svc.ds.AppConfig(ctx)
+	if err != nil {
+		return osqueryError{message: "ingest detail query: " + err.Error()}
+	}
+
+	detailQueries := osquery_utils.GetDetailQueries(config, svc.config)
+	query, ok := detailQueries[name]
+	if !ok {
+		return osqueryError{message: "unknown detail query " + name}
+	}
+
+	if query.IngestFunc != nil {
+		err = query.IngestFunc(svc.logger, host, rows)
+		if err != nil {
+			return osqueryError{
+				message: fmt.Sprintf("ingesting query %s: %s", name, err.Error()),
+			}
+		}
+	}
+
+	return nil
+}
+
+// filterPolicyResults filters out policies that aren't configured for webhook automation.
+func filterPolicyResults(incoming map[uint]*bool, webhookPolicies []uint) map[uint]*bool {
+	wp := make(map[uint]struct{})
+	for _, policyID := range webhookPolicies {
+		wp[policyID] = struct{}{}
+	}
+	filtered := make(map[uint]*bool)
+	for policyID, passes := range incoming {
+		if _, ok := wp[policyID]; !ok {
+			continue
+		}
+		filtered[policyID] = passes
+	}
+	return filtered
+}
+
+func (svc *Service) registerFlippedPolicies(ctx context.Context, hostID uint, hostname string, newFailing, newPassing []uint) error {
+	host := fleet.PolicySetHost{
+		ID:       hostID,
+		Hostname: hostname,
+	}
+	for _, policyID := range newFailing {
+		if err := svc.failingPolicySet.AddHost(policyID, host); err != nil {
+			return err
+		}
+	}
+	for _, policyID := range newPassing {
+		if err := svc.failingPolicySet.RemoveHosts(policyID, []fleet.PolicySetHost{host}); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func (svc *Service) maybeDebugHost(
+	ctx context.Context,
+	host *fleet.Host,
+	results fleet.OsqueryDistributedQueryResults,
+	statuses map[string]fleet.OsqueryStatus,
+	messages map[string]string,
+) {
+	if svc.debugEnabledForHost(ctx, host.ID) {
+		hlogger := log.With(svc.logger, "host-id", host.ID)
+
+		logJSON(hlogger, host, "host")
+		logJSON(hlogger, results, "results")
+		logJSON(hlogger, statuses, "statuses")
+		logJSON(hlogger, messages, "messages")
+	}
 }
